Split "menu" package into sub packages (menu.browser, menu.context). (#1201)
