+++ app/src/main/java/org/mozilla/focus/fragment/BrowserFragment.java
@@ -48,8 +48,8 @@ import org.mozilla.focus.architecture.NonNullObserver;
 import org.mozilla.focus.broadcastreceiver.DownloadBroadcastReceiver;
 import org.mozilla.focus.customtabs.CustomTabConfig;
 import org.mozilla.focus.locale.LocaleAwareAppCompatActivity;
-import org.mozilla.focus.menu.BrowserMenu;
-import org.mozilla.focus.menu.WebContextMenu;
+import org.mozilla.focus.menu.browser.BrowserMenu;
+import org.mozilla.focus.menu.context.WebContextMenu;
 import org.mozilla.focus.open.OpenWithFragment;
 import org.mozilla.focus.session.NullSession;
 import org.mozilla.focus.session.Session;
