+++ app/src/main/java/org/mozilla/focus/menu/browser/CustomTabMenuItemViewHolder.java
@@ -1,4 +1,4 @@
-package org.mozilla.focus.menu;
+package org.mozilla.focus.menu.browser;
 
 import android.app.PendingIntent;
 import android.content.Intent;
