+++ app/src/main/java/org/mozilla/focus/menu/browser/NavigationItemViewHolder.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.menu;
+package org.mozilla.focus.menu.browser;
 
 import android.view.View;
