+++ app/src/main/java/org/mozilla/focus/menu/context/WebContextMenu.java
@@ -2,7 +2,7 @@
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-package org.mozilla.focus.menu;
+package org.mozilla.focus.menu.context;
 
 import android.app.Dialog;
 import android.content.ClipData;
